agent-service/requirements.txt
livekit-agents
livekit-plugins-deepgram
livekit-plugins-google
livekit-plugins-silero
python-dotenv
aiohttp
